+++ scripts/test/DB/Ping.t
@@ -1,0 +1,60 @@
+# --
+# Ping.t - database tests
+# Copyright (C) 2001-2015 OTRS AG, http://otrs.com/
+# --
+# This software comes with ABSOLUTELY NO WARRANTY. For details, see
+# the enclosed file COPYING for license information (AGPL). If you
+# did not receive this file, see http://www.gnu.org/licenses/agpl.txt.
+# --
+
+use strict;
+use warnings;
+use utf8;
+
+use vars (qw($Self));
+
+# get needed objects
+my $DBObject = $Kernel::OM->Get('Kernel::System::DB');
+
+my @TestsMofifiers = (
+    {
+        Name    => 'Normal',
+        Command => '',
+        Success => 1,
+    },
+    {
+        Name    => 'Diconnected',
+        Command => 'Disconnect',
+        Success => 0,
+    },
+    {
+        Name    => 'Connected back',
+        Command => 'Connect',
+        Success => 1,
+    },
+);
+
+for my $TestCase (@TestsMofifiers) {
+
+    my $Command = $TestCase->{Command} || '';
+    if ($Command) {
+        $DBObject->$Command();
+    }
+
+    my $Success = $DBObject->Ping();
+
+    if ( $TestCase->{Success} ) {
+        $Self->True(
+            $Success,
+            "$TestCase->{Name} - Ping() with true",
+        );
+    }
+    else {
+        $Self->False(
+            $Success,
+            "$TestCase->{Name} - Ping() with false",
+        );
+    }
+}
+
+1;